requirements.txt
# ───── Web framework ─────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0

# ───── Google GenAI SDK (AI Studio + Vertex AI) ─────
google-genai>=1.10.0

# ───── Google Cloud TTS (cuando USAR_VERTEX=true) ─────
google-cloud-texttospeech>=2.16.0

# ───── Autenticación Google Cloud ─────
google-auth>=2.28.0

# ───── TTS: Edge (gratis) ─────
edge-tts>=7.0.0
# pydub>=0.25.1            # opcional si no tienes ffmpeg

# ───── Audio ─────
sounddevice==0.5.1
numpy==2.1.3

# VAD (más preciso que RMS, opcional pero recomendado)
webrtcvad-wheels>=2.0.14

# ───── Visión ─────
mediapipe==0.10.14
opencv-python==4.10.0.84

# ───── Screenshot ─────
mss>=9.0.1                  # captura multi-monitor
Pillow>=10.0.0              # fallback de captura + JPEG

# ───── Memoria / RAG ─────
chromadb==0.5.23

# ───── WebSockets ─────
websockets==14.1

# ───── Tests ─────
pytest>=8.0.0
pytest-asyncio>=0.23.0
